Move distributionTesting to plugin
