+++ buildSrc/subprojects/cleanup/cleanup.gradle.kts
@@ -2,6 +2,5 @@ apply { plugin("org.gradle.kotlin.kotlin-dsl")}
 
 dependencies {
     implementation(project(":configuration"))
-    implementation(project(":testing"))
     implementation(project(":kotlinDsl"))
 }
