+++ buildSrc/subprojects/testing/testing.gradle.kts
@@ -1,10 +1,21 @@
+plugins {
+    `java-gradle-plugin`
+}
+
 apply { plugin("org.gradle.kotlin.kotlin-dsl") }
 
 dependencies {
-    compile(project(":build"))
-    compile(project(":kotlinDsl"))
-    compile("org.codehaus.groovy.modules.http-builder:http-builder:0.7.2") {
-        // Xerces on the runtime classpath is breaking some of our doc tasks
-        exclude(group = "xerces")
+    implementation(project(":build"))
+    implementation(project(":kotlinDsl"))
+    implementation(project(":configuration"))
+    implementation(project(":cleanup"))
+}
+
+gradlePlugin {
+    (plugins) {
+        "distributionTesting" {
+            id = "gradlebuild.distribution-testing"
+            implementationClass = "org.gradle.testing.DistributionTestingPlugin"
+        }
     }
 }
